refactor [5/5]: Compute returnOnInvestment data on loading the cached data
So that it's already available in `studyData`

I tried to sort out the variables (sub key `eco`, subfolder `eco`) to layout the file structure and data structure when we'll compute more things here.

Note that it does add computation whenever we load (all) studies.
but because the volume of data is still small, it should'nt cause any slowdown
