+++ src/components/study/economic-growth/ReturnOnInvestment.vue
@@ -52,11 +52,10 @@ const handleDataChartSeriesClick = (event) => {
 }
 
 const { prettyAmount, convertAmount } = useCurrencyUtils(props)
-const returnOnInvestmentData = computed(() => buildReturnOnInvestmentData(props.studyData));
 
 const populatedBarChartData = computed(() => {
   let tooltip = {}
-  const items = returnOnInvestmentData.value.stages.map((stage) => {
+  const items = props.studyData.metrics.eco.returnOnInvestment.stages.map((stage) => {
     if (!stage.netOperatingProfits) {
       return null
     }
@@ -84,7 +83,7 @@ const populatedBarChartData = computed(() => {
 })
 
 const currentStageReturnOnInvestmentData = computed(() => {
-  const currentStage = returnOnInvestmentData.value.stages.find(stage => stage.name === selectedStage.value);
+  const currentStage = props.studyData.metrics.eco.returnOnInvestment.stages.find(stage => stage.name === selectedStage.value);
   const currentStageActors = currentStage.actors;
   const tooltip = {}
   const items = currentStageActors.map((actor) => {
@@ -99,44 +98,6 @@ const currentStageReturnOnInvestmentData = computed(() => {
   return getMiniBarChart(items, tooltip, formatPercent, getColor(selectedStage.value))
 })
 
-function buildReturnOnInvestmentData(studyData) {
-  const stages = studyData.ecoData.stages;
-  const actors = studyData.ecoData.actors;
-
-  return {
-    stages: stages.map(stage => buildStageReturnOnInvestmentData(stage, actors))
-  }
-}
-function buildStageReturnOnInvestmentData(stage, actors) {
-  const stageActors = actors.filter((actor) => actor.stage === stage.name).map(buildActorReturnOnInvestmentData);
-
-  const netOperatingProfits = _.sumBy(stageActors, "netOperatingProfits");
-  const totalCosts = _.sumBy(stageActors, "totalCosts");
-
-  return {
-    name: stage.name,
-    actors: stageActors,
-    netOperatingProfits,
-    totalCosts,
-    benefitCostRatio: netOperatingProfits / totalCosts,
-  }
-}
-
-function buildActorReturnOnInvestmentData(actor) {
-  const netOperatingProfits = actor.netOperatingProfit || 0;
-  let totalCosts = actor.totalCosts;
-
-  if (actor.stage === 'Producers') {
-    totalCosts += netOperatingProfits;
-  }
-  return {
-    name: actor.name,
-    stage: actor.stage,
-    netOperatingProfits,
-    totalCosts,
-    benefitCostRatio: netOperatingProfits / totalCosts,
-  };
-}
 </script>
 
 <style scoped lang="scss">
