+++ src/utils/data/metrics/index.js
@@ -1,0 +1,9 @@
+import { buildReturnOnInvestmentData } from "./eco/returnOnInvestment";
+
+export function computeMetrics(studyData) {
+  return {
+    eco: {
+      returnOnInvestment: buildReturnOnInvestmentData(studyData)
+    }
+  }
+}
